Added login screen and communication between android and Authservice
