+++ mobileapp/.idea/misc.xml
@@ -5,6 +5,7 @@
       <map>
         <entry key="..\:/Avans/Jaar 3/Project Everyware/Code/VirtualAssistant/mobile-app/app/src/main/res/layout/activity_main.xml" value="0.358695652173913" />
         <entry key="app/src/main/res/drawable/mic_inactive.xml" value="0.1" />
+        <entry key="app/src/main/res/layout/activity_assistant.xml" value="0.3458333333333333" />
         <entry key="app/src/main/res/layout/activity_chat.xml" value="0.33897435897435896" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.24069478908188585" />
         <entry key="app/src/main/res/layout/frame_chat_me.xml" value="0.1" />
