+++ mobileapp/app/src/main/java/com/infosupport/virtualassistent/bot/models/LoginAsyncResponse.java
@@ -1,0 +1,5 @@
+package com.infosupport.virtualassistent.bot.models;
+
+public interface LoginAsyncResponse {
+    void processFinished(boolean correctCredentials, String response);
+}
