+++ mobileapp/app/src/main/java/com/infosupport/virtualassistent/bot/websocket/BotWebSocketClient.java
@@ -1,6 +1,6 @@
 package com.infosupport.virtualassistent.bot.websocket;
 
-import com.infosupport.virtualassistent.MainActivity;
+import com.infosupport.virtualassistent.AssistantActivity;
 import com.infosupport.virtualassistent.bot.models.Schedule;
 
 import org.java_websocket.client.WebSocketClient;
@@ -13,12 +13,12 @@ import java.lang.ref.WeakReference;
 import java.net.URI;
 
 public class BotWebSocketClient extends WebSocketClient {
-    private final WeakReference<MainActivity> activityRef;
+    private final WeakReference<AssistantActivity> activityRef;
     private String lastMessage = "";
 
-    public BotWebSocketClient(URI serverURI, MainActivity activity) {
+    public BotWebSocketClient(URI serverURI, AssistantActivity activity) {
         super(serverURI);
-        activityRef = new WeakReference<MainActivity>(activity);
+        activityRef = new WeakReference<AssistantActivity>(activity);
     }
 
     @Override
