+++ mobileapp/app/src/main/java/com/infosupport/virtualassistent/receivers/DetectionResultReceiver.java
@@ -2,16 +2,16 @@ package com.infosupport.virtualassistent.receivers;
 
 import android.media.MediaPlayer;
 
-import com.infosupport.virtualassistent.MainActivity;
+import com.infosupport.virtualassistent.AssistantActivity;
 import com.infosupport.virtualassistent.R;
 
 import java.lang.ref.WeakReference;
 
 public class DetectionResultReceiver implements SpeechResultReceiver.ResultReceiverCallBack<String> {
-    private final WeakReference<MainActivity> activityRef;
+    private final WeakReference<AssistantActivity> activityRef;
 
-    public DetectionResultReceiver(MainActivity activity) {
-        activityRef = new WeakReference<MainActivity>(activity);
+    public DetectionResultReceiver(AssistantActivity activity) {
+        activityRef = new WeakReference<AssistantActivity>(activity);
     }
 
     @Override
