+++ mobileapp/app/src/main/java/com/infosupport/virtualassistent/receivers/RecognizeSpeechResultReceiver.java
@@ -1,14 +1,14 @@
 package com.infosupport.virtualassistent.receivers;
 
-import com.infosupport.virtualassistent.MainActivity;
+import com.infosupport.virtualassistent.AssistantActivity;
 
 import java.lang.ref.WeakReference;
 
 public class RecognizeSpeechResultReceiver implements SpeechResultReceiver.ResultReceiverCallBack<String> {
-    private final WeakReference<MainActivity> activityRef;
+    private final WeakReference<AssistantActivity> activityRef;
 
-    public RecognizeSpeechResultReceiver(MainActivity activity) {
-        activityRef = new WeakReference<MainActivity>(activity);
+    public RecognizeSpeechResultReceiver(AssistantActivity activity) {
+        activityRef = new WeakReference<AssistantActivity>(activity);
     }
 
     @Override
