+++ mobileapp/app/src/main/res/layout/activity_assistant.xml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+xmlns:app="http://schemas.android.com/apk/res-auto"
+xmlns:tools="http://schemas.android.com/tools"
+android:layout_width="match_parent"
+android:layout_height="match_parent">
+
+<androidx.recyclerview.widget.RecyclerView
+    android:id="@+id/recycler_gchat"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:layout_marginBottom="16dp"
+    app:layout_constrainedHeight="true"
+    app:layout_constraintBottom_toTopOf="@+id/text_indicator" />
+
+<TextView
+    android:id="@+id/text_indicator"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_marginStart="16dp"
+    android:visibility="gone"
+    app:layout_constraintBottom_toTopOf="@id/view"
+    app:layout_constraintStart_toStartOf="parent" />
+
+<View
+    android:id="@+id/view"
+    android:layout_width="match_parent"
+    android:layout_height="1dp"
+    android:layout_marginBottom="10dp"
+    android:background="#e0e0e0"
+    app:layout_constraintBottom_toTopOf="@+id/fab" />
+
+<com.google.android.material.floatingactionbutton.FloatingActionButton
+    android:id="@+id/fab"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_gravity="bottom|center"
+    android:layout_marginBottom="10dp"
+    android:contentDescription="@string/listen_button"
+    android:src="@drawable/mic_inactive"
+    app:backgroundTint="@color/design_default_color_primary"
+    app:rippleColor="@color/design_default_color_background"
+    app:fabCustomSize="92dp"
+    app:tint="@android:color/white"
+    app:maxImageSize="60dp"
+    app:layout_constraintBottom_toBottomOf="parent"
+    app:layout_constraintEnd_toEndOf="parent"
+    app:layout_constraintHorizontal_bias="0.5"
+    app:layout_constraintStart_toStartOf="parent" />
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
