+++ mobileapp/app/src/main/res/values/strings.xml
@@ -5,4 +5,8 @@
     <string name="assistant_name">Assistent</string>
     <string name="channel_name">WWE_Notification</string>
     <string name="channel_description">A notification that shows whenever the Wake Word Engine is active</string>
+    <string name="login_form_title">Welkom!</string>
+    <string name="login_form_username">Gebruikersnaam</string>
+    <string name="login_form_password">Wachtwoord</string>
+    <string name="login_form_login">Inloggen</string>
 </resources>
